successful transimit

+++ inc/hw_config.h
@@ -62,8 +62,8 @@ void USB_Interrupts_Config(void);
 void USB_Cable_Config (FunctionalState NewState);
 void Get_SerialNum(void);
 void LCD_Control(void);
-uint32_t CDC_Send_DATA (uint8_t *ptrBuffer, uint8_t Send_length);
-uint32_t CDC_Receive_DATA(void);
+uint8_t CDC_Send_DATA (uint8_t *ptrBuffer, uint8_t Send_length);
+uint8_t CDC_Receive_DATA(void);
 /* External variables --------------------------------------------------------*/
 
 #endif  /*__HW_CONFIG_H*/
